Add files via upload
Check the additions to see what I have added. Most of them are done and need to be tested. The multiply method and the power method I haven't done yet, and all of the new stuff needs to be tested in main. I have a decent skeleton in for the method in Matrix that changes into a SparseMatrix. Let me know if you need help, taking a break to rest for a bit.